pyproject.toml
[project]
name = "aegis-discord-bot"
version = "0.1.0"
description = "A Vortex-inspired moderation bot for Discord, built in Python."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiohttp>=3.10.0,<4.0.0",
  "aiosqlite>=0.20.0,<1.0.0",
  "discord.py>=2.6.0,<2.7.0",
  "PyNaCl>=1.5.0,<2.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
]

[project.optional-dependencies]
docs = [
  "mkdocs>=1.6.0,<2.0.0",
  "mkdocs-material>=9.5.0,<10.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["aegis"]

[tool.ruff]
line-length = 100
target-version = "py311"
